+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Deprecate `create_ambiorix`: moving to [ambiorix.generator](https://github.com/JohnCoene/ambiorix.generator) package.
 - Deprecate `add_template`: moving to [ambiorix.generator](https://github.com/JohnCoene/ambiorix.generator) package.
+- Added `all` method to define route and handler for all methods `GET`, `POST`, `PUT`, `DELETE`, and `PATCH`.
 
 # ambiorix 1.0.0